ERS-240 Moved duplicated methods from BaseActivityViewMapper and GearCharacteristicsMapper to BaseMapper. Removed some unused mapping methods.
