+++ service/src/main/java/eu/europa/ec/fisheries/ers/service/mapper/GearCharacteristicsMapper.java
@@ -12,13 +12,9 @@ details. You should have received a copy of the GNU General Public License along
 package eu.europa.ec.fisheries.ers.service.mapper;
 
 import eu.europa.ec.fisheries.ers.fa.entities.FishingGearEntity;
-import eu.europa.ec.fisheries.ers.fa.entities.FishingGearRoleEntity;
 import eu.europa.ec.fisheries.ers.fa.entities.GearCharacteristicEntity;
 import eu.europa.ec.fisheries.ers.service.dto.view.GearDto;
 import eu.europa.ec.fisheries.ers.service.util.CustomBigDecimal;
-import eu.europa.ec.fisheries.ers.service.util.Utils;
-import org.apache.commons.collections.CollectionUtils;
-import org.apache.commons.lang3.StringUtils;
 import org.mapstruct.InheritInverseConfiguration;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
@@ -27,24 +23,10 @@ import org.mapstruct.ReportingPolicy;
 import org.mapstruct.factory.Mappers;
 import un.unece.uncefact.data.standard.reusableaggregatebusinessinformationentity._20.GearCharacteristic;
 
-import java.util.Objects;
-import java.util.Set;
-
-import static eu.europa.ec.fisheries.ers.service.mapper.view.base.ViewConstants.GEAR_CHARAC_Q_CODE_C62;
-import static eu.europa.ec.fisheries.ers.service.mapper.view.base.ViewConstants.GEAR_CHARAC_TYPE_CODE_GD;
-import static eu.europa.ec.fisheries.ers.service.mapper.view.base.ViewConstants.GEAR_CHARAC_TYPE_CODE_GM;
-import static eu.europa.ec.fisheries.ers.service.mapper.view.base.ViewConstants.GEAR_CHARAC_TYPE_CODE_GN;
-import static eu.europa.ec.fisheries.ers.service.mapper.view.base.ViewConstants.GEAR_CHARAC_TYPE_CODE_HE;
-import static eu.europa.ec.fisheries.ers.service.mapper.view.base.ViewConstants.GEAR_CHARAC_TYPE_CODE_ME;
-import static eu.europa.ec.fisheries.ers.service.mapper.view.base.ViewConstants.GEAR_CHARAC_TYPE_CODE_NI;
-import static eu.europa.ec.fisheries.ers.service.mapper.view.base.ViewConstants.GEAR_CHARAC_TYPE_CODE_NL;
-import static eu.europa.ec.fisheries.ers.service.mapper.view.base.ViewConstants.GEAR_CHARAC_TYPE_CODE_NN;
-import static eu.europa.ec.fisheries.ers.service.mapper.view.base.ViewConstants.GEAR_CHARAC_TYPE_CODE_QG;
-
 @Mapper(imports = BaseMapper.class,
         uses = CustomBigDecimal.class,
         unmappedTargetPolicy = ReportingPolicy.ERROR)
-public abstract class GearCharacteristicsMapper {
+public abstract class GearCharacteristicsMapper extends BaseMapper {
 
     public static final GearCharacteristicsMapper INSTANCE = Mappers.getMapper(GearCharacteristicsMapper.class);
 
@@ -80,57 +62,5 @@ public abstract class GearCharacteristicsMapper {
         gearDto.setType(fishingGearEntity.getTypeCode());
         fillRoleAndCharacteristics(gearDto, fishingGearEntity);
         return gearDto;
-    }
-
-    private void fillRoleAndCharacteristics(GearDto gearDto, FishingGearEntity gearEntity) {
-        Set<FishingGearRoleEntity> fishingGearRole = gearEntity.getFishingGearRole();
-        if (CollectionUtils.isNotEmpty(fishingGearRole)) {
-            FishingGearRoleEntity role = fishingGearRole.iterator().next();
-            gearDto.setRole(role.getRoleCode());
-        }
-        Set<GearCharacteristicEntity> gearCharacteristics = gearEntity.getGearCharacteristics();
-        for (GearCharacteristicEntity charac : Utils.safeIterable(gearCharacteristics)) {
-            fillCharacteristicField(charac, gearDto);
-        }
-    }
-
-    private void fillCharacteristicField(GearCharacteristicEntity charac, GearDto gearDto) {
-
-        String quantityOnly = charac.getValueMeasure() != null ? charac.getValueMeasure().toString() : StringUtils.EMPTY;
-        String quantityWithUnit = quantityOnly + charac.getValueMeasureUnitCode();
-
-        switch (charac.getTypeCode()) {
-            case GEAR_CHARAC_TYPE_CODE_ME:
-                gearDto.setMeshSize(quantityWithUnit);
-                break;
-            case GEAR_CHARAC_TYPE_CODE_GM:
-                gearDto.setLengthWidth(quantityWithUnit);
-                break;
-            case GEAR_CHARAC_TYPE_CODE_GN:
-                gearDto.setNumberOfGears(Integer.parseInt(quantityOnly));
-                break;
-            case GEAR_CHARAC_TYPE_CODE_HE:
-                gearDto.setHeight(quantityWithUnit);
-                break;
-            case GEAR_CHARAC_TYPE_CODE_NI:
-                gearDto.setNrOfLines(quantityWithUnit);
-                break;
-            case GEAR_CHARAC_TYPE_CODE_NN:
-                gearDto.setNrOfNets(quantityWithUnit);
-                break;
-            case GEAR_CHARAC_TYPE_CODE_NL:
-                gearDto.setNominalLengthOfNet(quantityWithUnit);
-                break;
-            case GEAR_CHARAC_TYPE_CODE_QG:
-                if (!Objects.equals(charac.getValueQuantityCode(), GEAR_CHARAC_Q_CODE_C62)) {
-                    gearDto.setQuantity(quantityWithUnit);
-                }
-                break;
-            case GEAR_CHARAC_TYPE_CODE_GD:
-                gearDto.setDescription(charac.getDescription());
-                break;
-            default:
-                break;
-        }
     }
 }
